Add metadata content component
Applies to both Blog and Comment.
Handling the timestamp is a little ugly, since it references different fields for the two types of content. If I wanted to extend the functionality to more types, I would need to genericize it.

This also means timestamps have been added to Blog component.